Create Update-AXVertexTable.ps1
There was a Vertex tax setting in General Ledger in AX that was not updated.  So when the refresh script runs, it captures the config tables and overlays those settings back after the databases are restored from prod, effectively erasing the new change implemented in Prod.  This script queries each ax server in the ComputerList.txt file, check for the missing row, and either updates or skips it with console feedback.