Create Get-TFSUserGroupMembership.ps1
I was asked to help the security auditors list the users that have access and their roles within TFS 2013.  This seems to require 3rd party tools, but what I was able to do is dump the tfs dbo.tbl_Identities table with headers to a csv file and use PowerShell to ingest the file as an array of custom objects (an object of reach line of the csv), then use the username to query AD to see if the account is enabled (true/false or not found), and then call a function to see if the user is a member of any of the elements in an array (list of groups) I saw within the TFS admin console.  Then it adds dynamically those items to each custom object (user object).  The result is array of custom objects with the extra elements (enabled, and each AD Group specified in the array) that can be exported to a csv for further processing.   The csv file columns are the member items from the custom object.  Each column can then be filtered (i.e. by Enabled users, or specific AD Group column)