Create Get-AWS_IPAddresses.PS1
Request from Sec dude.
They are pulling a list of IP addresses from AWS, filtering for US-East1 and outputting a file.  The target location is a UNC path that will be some web server.  The web server will feed a firewall box.  Beyond that I'm not sure.  I added the parameters for easy edit and the try-catch error handling.